Move model manager and model feedback under modeling folder
