+++ eis_qgis_plugin/wizard/modeling/model_manager.py
